chore(refactor): factor out o- & e-call api mods
